管理员页面已实现

+++ freetalk/postbar/models.py
@@ -2,13 +2,14 @@ from django.db import models
 from django.contrib.auth.models import User
 from django.utils import timezone
 import os
+from freetalk.settings import MEDIA_ROOT
 # Create your models here.
 # type     0: normal user|1: admin |2: superadmin
 # status   0: normal     |1: forbid|2: hidden
 # resptype 0: normal     |1: the resp of resp
 def user_directory_path(instance, filename):
-	if os.path.isfile('E:/Works/github/BackEndProjectFreetalk/freetalk/media/upload/' + str(instance.user.id)) == True:
-		os.remove('E:/Works/github/BackEndProjectFreetalk/freetalk/media/upload/' + str(instance.user.id))
+	if os.path.isfile(MEDIA_ROOT + '/upload/' + str(instance.user.id)) == True:
+		os.remove(MEDIA_ROOT + '/upload/' + str(instance.user.id))
 	return 'upload/' + str(instance.user.id)
 
 class TKuser(models.Model):
@@ -35,45 +36,63 @@ class TKuser(models.Model):
 		self.img = newImg
 		self.save()
 
-	def modifyStatus(self, usrId, newStatus):
-		q = User.objects.filter(id = usrId)[0]
-		q.tkuser.usrStatus = newStatus
-		q.tkuser.save()
+	def modifyStatus(self, newStatus):
+		self.usrStatus = newStatus
+		self.save()
 
-	def modifyPermission(self, usrId, newPermission):
-		q = User.objects.filter(id = usrId)[0]
-		q.tkuser.usrType = newPermission
-		q.tkuser.save()
+	def modifyPermission(self, newPermission):
+		self.usrType = newPermission
+		self.save()
 
 	def getPost(self):
 		return TKpost.objects.filter(user = self.user)
 
-	def newPost(self):
-		pass
+	def newPost(self, title, content, img, attachment, classTag, keyword):
+		q = TKpost(title = title, content = content, img = img, attachment = attachment, classTag = classTag, keyword = keyword, user = self.user)
+		q.save()
+		self.numPost = self.numPost + 1
 
-	def deletePost():
-		pass
+	def deletePost(self, postId):
+		q = TKpost.objects.filter(id = postId)
+		if q:
+			q = q[0]
+			q.delete()
+			q.user.tkuser.numPost = q.user.tkuser.numPost - 1
 
-	def upvotePost():
-		pass
+	def upvotePost(self, postId):
+		q = TKpost.objects.filter(id = postId)
+		if q:
+			q = q[0]
+			q.score = q.score + 1
+			q.user.tkuser.numScore = q.user.tkuser.numScore + 1
 
-	def giveCoinPost():
-		pass
+	def giveCoinPost(self, postId):
+		q = TKpost.objects.filter(id = postId)
+		if q:
+			q = q[0]
+			q.coin = q.coin + 1
+			q.user.tkuser.numCoin = q.user.tkuser.numCoin + 1
 
-	def newResp():
-		pass
+	def newResp(self, respType, content, postId, respId, hostId):
+		q = None
+		p = TKpost.objects.filter(id = postId)
+		p = p[0] if p else None
+		q = TKresponse(respType = respType, content = content, user = self.user, post = p, respId = respId, hostId = hostId)
+		q.save()
 
-	def deleteResp():
-		pass
+	def deleteResp(self, respId):
+		q = TKresponse.objects.filter(id = respId)
+		if q:
+			q = q[0]
+			TKresponse.objects.filter(respId = respId).delete()
+			q.delete()
 
 	def upvoteResp():
-		pass
-
-	def addClassTag(newClassTag):
-		pass
-
-	def delClassTag(classTag):
-		pass
+		q = TKpost.objects.filter(id = postId)
+		if q:
+			q = q[0]
+			q.score = q.score + 1
+			q.user.tkuser.numScore = q.user.tkuser.numScore + 1
 
 
 class TKpost(models.Model):
@@ -115,22 +134,61 @@ class TKclassTag(models.Model):
 	classTagName = models.CharField(max_length = 100)
 
 class TKhomepage:
+	def newAdmin():
+		if User.objects.filter(username = 'admin'):
+			return
+		q = User.objects.create_user('admin', '', 'abcdefgh')
+		q.is_staff = True
+		q.is_superuser = True
+		q.save()
+		u = TKuser(nickname = 'admin', pwdQuestion = 'Are you admin?', pwdAnswer = 'Maybe', user = q)
+		u.usrType = 2
+		u.save()
+
 	def newUser(username, password, email, nickname, pwdQuestion, pwdAnswer):
 		q = User.objects.create_user(username, email, password)
 		q.save()
 		u = TKuser(nickname = nickname, pwdQuestion = pwdQuestion, pwdAnswer = pwdAnswer, user = q)
 		u.save()
 
-	def deleteUser(usrId):
-		q = User.objects.filter(id = usrId)[0]
-		q.delete()
+	def deleteUser(username):
+		q = User.objects.filter(username = username)
+		if q:
+			q = q[0]
+			if os.path.isfile(MEDIA_ROOT + '/upload/' + str(q.id)) == True:
+				os.remove(MEDIA_ROOT + '/upload/' + str(q.id))
+			q = User.objects.filter(username = username)
+			q.delete()
+
+
+	def addClassTag(newClassTag):
+		q = TKclassTag(classTagName = newClassTag)
+		q.save()
+
+	def delClassTag(classTag):
+		q = TKclassTag.objects.filter(classTagName = classTag)
+		if q:
+			q = q[0]
+			q.delete()
 
 	def searchUsrByNickname(nickname):
-		pass
+		q = tkuser.objects.filter(nickname = nickname)
+		return [p.user for p in q]
 
 	def searchUsrById(usrId):
-		pass
-		
+		q = User.objects.filter(id = usrId)
+		if q:
+			return q[0]
+		else:
+			return None
+	
+	def searchUsrByName(name):
+		q = User.objects.filter(username = name)
+		if q:
+			return q[0]
+		else:
+			return None
+
 	def searchPostByUsrId(usrId):
 		return TKpost.objects.filter(user = User.objects.filter(id = usrId)[0])
